feat: change to atomic design
