+++ .env
@@ -1,6 +1,6 @@
 VITE_API_BASE_URL=https://coipekj2sl.execute-api.ap-northeast-1.amazonaws.com/dev
 VITE_CHAT_API_URL=https://67hnjuna66.execute-api.ap-northeast-1.amazonaws.com/prd-1
-VITE_ACCENT_COLOR=orange
+VITE_ACCENT_COLOR=red
 VITE_SHOW_NAVIGATION_HEADER=true
 VITE_SHOW_TIMESTAMP=true
 VITE_FAQ_ICONS={"category1":{"type":"component","value":"ImageIcon"},"category2":{"type":"lucide","value":"Users"},"category3":{"type":"lucide","value":"Trophy"},"category4":{"type":"lucide","value":"FileText"},"other":{"type":"lucide","value":"MoreHorizontal"}}
