+++ src/components/atoms/Icon/index.ts
@@ -1,0 +1,2 @@
+export { Icon } from './Icon';
+export { Icon as DynamicIcon } from './Icon'; // 기존 이름 호환성을 위해
