+++ src/components/atoms/Typography/index.ts
@@ -1,0 +1,2 @@
+export { TypewriterText } from './Typography';
+export { TypewriterText as default } from './Typography'; // 기존 default export 호환성을 위해
