+++ src/components/atoms/index.ts
@@ -1,0 +1,3 @@
+export * from './Button';
+export * from './Icon';
+export * from './Typography';
