+++ src/components/molecules/ChatBubble/index.ts
@@ -1,0 +1,1 @@
+export { ChatBubble } from './ChatBubble';
