+++ src/components/molecules/InputField/index.ts
@@ -1,0 +1,1 @@
+export { InputField } from './InputField';
