+++ src/components/molecules/UserAvatar/index.ts
@@ -1,0 +1,1 @@
+export { UserAvatar } from './UserAvatar';
