+++ src/components/organisms/ChatInput/index.ts
@@ -1,0 +1,2 @@
+export { ChatInput } from './ChatInput';
+export { ChatInput as default } from './ChatInput';
