+++ src/components/organisms/ChatMessage/index.ts
@@ -1,0 +1,2 @@
+export { ChatMessage } from './ChatMessage';
+export { ChatMessage as default } from './ChatMessage';
