+++ src/components/organisms/FAQCategory/index.ts
@@ -1,0 +1,3 @@
+export { FAQCategory } from './FAQCategory';
+export { FAQCategory as default } from './FAQCategory';
+export type { FAQCategoryItem } from './FAQCategory';
