+++ src/components/organisms/QuickReply/index.ts
@@ -1,0 +1,3 @@
+export { QuickReply } from './QuickReply';
+export { QuickReply as default } from './QuickReply';
+export type { QuickReplyOption } from './QuickReply';
