+++ src/components/organisms/TopQuestions/index.ts
@@ -1,0 +1,2 @@
+export { TopQuestions } from './TopQuestions';
+export { TopQuestions as default } from './TopQuestions';
