+++ src/components/organisms/index.ts
@@ -1,0 +1,6 @@
+export * from './ChatInput';
+export * from './ChatMessage';
+export * from './FAQCategory';
+export * from './QuickReply';
+export * from './TopQuestions';
+export * from './NavigationHeader';
