+++ src/components/templates/ChatLayout/index.ts
@@ -1,0 +1,1 @@
+export { ChatLayout } from './ChatLayout';
